Added change_notes.kt script to buildSrc and use it in 251 plugin config.
