+++ hot_preview_render_1_8/build.gradle.kts
@@ -1,7 +1,7 @@
 plugins {
     kotlin("jvm")
     kotlin("plugin.compose") version Versions.kotlin
-    id("org.jetbrains.compose") version "1.8.0"
+    id("org.jetbrains.compose") version "1.8.2"
     id("com.gradleup.shadow") version Versions.shadow
 }
