+++ intellij-plugin/core/build.gradle.kts
@@ -1,6 +1,3 @@
-import org.intellij.lang.annotations.Language
-import java.nio.file.Files
-import java.nio.file.Paths
 import java.util.*
 
 plugins {
@@ -22,43 +19,6 @@ repositories {
 
 val rootProperties = Properties()
 rootProject.file("gradle.properties").inputStream().use { rootProperties.load(it) }
-
-data class VersionInfo(
-  val version: String,
-  val items: MutableList<String> = mutableListOf()
-)
-
-@Language("HTML")
-val changeNotesText = run {
-  val releaseNotesFile = rootProject.file("RELEASE_NOTES_PLUGIN.md")
-  val versionModel = mutableListOf<VersionInfo>()
-  var currentVersion: VersionInfo? = null
-  var versionCount = 4
-  releaseNotesFile.readLines().forEach { line ->
-    if (line.startsWith("### ") && versionCount > 0) {
-      versionCount--
-      if (versionCount > 0) {
-        val version = line.substring(3).trim()
-        currentVersion = VersionInfo(version).also {
-          versionModel.add(it)
-        }
-      }
-    } else if (versionCount > 0 && line.startsWith("- ")) {
-      currentVersion?.items?.add(line.substring(2).trim())
-    }
-  }
-
-  val htmlBuilder = StringBuilder()
-  versionModel.forEach { versionInfo ->
-    htmlBuilder.append("<h3>${versionInfo.version}</h3>")
-    htmlBuilder.append("<ul>")
-    versionInfo.items.forEach { items ->
-      htmlBuilder.append("<li>$items</li>")
-    }
-    htmlBuilder.append("</ul>")
-  }
-  htmlBuilder.toString()
-}
 
 
 dependencies {
