+++ intellij-plugin/plugin-251/plugin.properties
@@ -1,5 +1,5 @@
 ijpTarget=251
 jdkLevel=21
-pluginSinceBuild=251
+pluginSinceBuild=251.1
 pluginUntilBuild=251.*
 #platformVersion comes from the root gradle.properties since this is the highest
